--- src/Request/search/index.js
@@ -1,3 +1,0 @@
-import request from '@/Request/base'
-import curry from '@/Utils/curry'
-export default curry(request.post, '/search')
