+++ src/components/Product/Partition.jsx
