+++ src/components/Slider.jsx
