+++ src/containers/Gifts.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+
+export default class GiftsContainer extends Component {
+  static propTypes = {
+  }
+
+  render() {
+    return (
+      <div>
+        
+      </div>
+    )
+  }
+}
