+++ src/network/search/index.js
@@ -1,0 +1,3 @@
+import request from '@/network/base'
+import curry from '@/utils/curry'
+export default curry(request.post, '/search')
